pyproject.toml
[project]
name = "binlog2sql-gui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "mysql-replication==0.44.0",
    "pymysql>=1.1.1",
    "pyside6>=6.9.0",
]
